Refactor submodules for SRP
