+++ qmtl/indicators/__init__.py
@@ -1,30 +1,5 @@
-"""Technical indicator nodes built on top of ``qmtl.sdk``.
+"""Technical indicator nodes built on top of ``qmtl.sdk``."""
 
-Modules here follow the Single Responsibility Principle and remain independent
-from other namespaces. Add tests under ``tests/`` when introducing new
-indicators.
-"""
+from .sma import sma
 
-from qmtl.sdk.node import Node
-from qmtl.sdk.node import StreamInput
-from qmtl.sdk.cache_view import CacheView
-
-
-def sma(source: Node, window: int, *, name: str | None = None) -> Node:
-    """Return a Node computing a simple moving average over ``window`` values."""
-
-    def compute(view: CacheView):
-        data = view[source][source.interval][-window:]
-        if not data:
-            return None
-        values = [v for _, v in data]
-        return sum(values) / len(values)
-
-    return Node(
-        input=source,
-        compute_fn=compute,
-        name=name or "sma",
-        interval=source.interval,
-        period=window,
-    )
-
+__all__ = ["sma"]
