+++ qmtl/streams/__init__.py
@@ -1,15 +1,5 @@
-"""StreamInput implementations for common data sources.
+"""StreamInput implementations for common data sources."""
 
-Each class keeps a narrow focus and does not depend on other namespaces.
-New streams must be accompanied by tests under ``tests/``.
-"""
+from .binance import BinanceBTCStream
 
-from qmtl.sdk.node import StreamInput
-
-
-class BinanceBTCStream(StreamInput):
-    """Placeholder StreamInput for Binance BTC trades."""
-
-    def __init__(self, *, interval: int, period: int) -> None:
-        super().__init__(tags=["binance", "btc"], interval=interval, period=period)
-
+__all__ = ["BinanceBTCStream"]
